+++ PopupDialog/Classes/PopupDialogButton.swift
@@ -44,8 +44,8 @@ open class PopupDialogButton: UIButton {
     
     /// The title color of the button
     @objc open dynamic var titleColor: UIColor? {
-        get { return self.titleColor(for: UIControlState()) }
-        set { setTitleColor(newValue, for: UIControlState()) }
+        get { return self.titleColor(for: UIControl.State()) }
+        set { setTitleColor(newValue, for: UIControl.State()) }
     }
 
     /// The background color of the button
@@ -120,7 +120,7 @@ open class PopupDialogButton: UIButton {
         super.init(frame: .zero)
 
         // Set the button title
-        setTitle(title, for: UIControlState())
+        setTitle(title, for: UIControl.State())
 
         self.dismissOnTap = dismissOnTap
 
@@ -137,7 +137,7 @@ open class PopupDialogButton: UIButton {
     open func setupView() {
 
         // Default appearance
-        setTitleColor(defaultTitleColor, for: UIControlState())
+        setTitleColor(defaultTitleColor, for: UIControl.State())
         titleLabel?.font              = defaultTitleFont
         backgroundColor               = defaultButtonColor
         separator.backgroundColor     = defaultSeparatorColor