+++ PopupDialog/Classes/UIView+Animations.swift
@@ -39,33 +39,33 @@ internal enum AnimationDirection {
 internal extension UIView {
 
     /// The key for the fade animation
-    internal var fadeKey: String { return "FadeAnimation" }
-    internal var shakeKey: String { return "ShakeAnimation" }
+    var fadeKey: String { return "FadeAnimation" }
+    var shakeKey: String { return "ShakeAnimation" }
 
-    internal func pv_fade(_ direction: AnimationDirection, _ value: Float, duration: CFTimeInterval = 0.08) {
+    func pv_fade(_ direction: AnimationDirection, _ value: Float, duration: CFTimeInterval = 0.08) {
         layer.removeAnimation(forKey: fadeKey)
         let animation = CABasicAnimation(keyPath: "opacity")
         animation.duration = duration
         animation.fromValue = layer.presentation()?.opacity
         layer.opacity = value
-        animation.fillMode = kCAFillModeForwards
+        animation.fillMode = CAMediaTimingFillMode.forwards
         layer.add(animation, forKey: fadeKey)
     }
 
-    internal func pv_layoutIfNeededAnimated(duration: CFTimeInterval = 0.08) {
-        UIView.animate(withDuration: duration, delay: 0, options: UIViewAnimationOptions(), animations: {
+    func pv_layoutIfNeededAnimated(duration: CFTimeInterval = 0.08) {
+        UIView.animate(withDuration: duration, delay: 0, options: UIView.AnimationOptions(), animations: {
             self.layoutIfNeeded()
         }, completion: nil)
     }
     
     // As found at https://gist.github.com/mourad-brahim/cf0bfe9bec5f33a6ea66#file-uiview-animations-swift-L9
     // Slightly modified
-    internal func pv_shake() {
+    func pv_shake() {
         layer.removeAnimation(forKey: shakeKey)
         let vals: [Double] = [-2, 2, -2, 2, 0]
         
         let translation = CAKeyframeAnimation(keyPath: "transform.translation.x")
-        translation.timingFunction = CAMediaTimingFunction(name: kCAMediaTimingFunctionLinear)
+        translation.timingFunction = CAMediaTimingFunction(name: CAMediaTimingFunctionName.linear)
         translation.values = vals
         
         let rotation = CAKeyframeAnimation(keyPath: "transform.rotation.z")