+++ Tests/SnapshotTests.swift
@@ -12,11 +12,11 @@ class SnapshotTests: FBSnapshotTestCase {
         super.setUp()
     
         recordMode = false
-        isDeviceAgnostic = true
+        fileNameOptions = [.device, .OS, .screenSize, .screenScale]
 
         let bundle = Bundle(for: self.classForCoder)
         let sb = UIStoryboard(name: "Main", bundle: bundle)
-        vc = sb.instantiateViewController(withIdentifier: "ViewController") as! ViewController
+        vc = sb.instantiateViewController(withIdentifier: "ViewController") as? ViewController
         XCTAssertNotNil(vc.view)
         
         window = UIWindow(frame: UIScreen.main.bounds)
@@ -31,21 +31,21 @@ class SnapshotTests: FBSnapshotTestCase {
     func testStandardDialogWithImage() {
         vc.showImageDialog(animated: false)
         let date = Date(timeIntervalSinceNow: 1)
-        RunLoop.current.run(mode: .defaultRunLoopMode, before: date)
+        RunLoop.current.run(mode: RunLoop.Mode.default, before: date)
         FBSnapshotVerifyView(window)
     }
     
     func testStandardDialogWithoutImage() {
         vc.showStandardDialog(animated: false)
         let date = Date(timeIntervalSinceNow: 1)
-        RunLoop.current.run(mode: .defaultRunLoopMode, before: date)
+        RunLoop.current.run(mode: RunLoop.Mode.default, before: date)
         FBSnapshotVerifyView(window)
     }
     
     func testCustomDialog() {
         vc.showCustomDialog(animated: false)
         let date = Date(timeIntervalSinceNow: 1)
-        RunLoop.current.run(mode: .defaultRunLoopMode, before: date)
+        RunLoop.current.run(mode: RunLoop.Mode.default, before: date)
         FBSnapshotVerifyView(window)
     }
 }