+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
-
+* **1.1.1** Updates dependencies to Swift 5
+* **1.1.0** Swift 5 support
+* **1.0.0** Pinned Swift version to 4.2<br>Dropped iOS 9 support as of moving to ios-snapshot-test-case
+* **0.9.2** Fixes crash when presenting dialog while app is inactive
+* **0.9.1** Fixes Carthage support
+* **0.9.0** Swift 4.2 support
 * **0.8.1** Added shadow appearance properties
 * **0.8.0** Separated tap and pan gesture dismissal
 * **0.7.1** Fixes Objective-C compatability<br>Improved Carthage handling
@@ -9,7 +14,7 @@
 * **0.6.0** Swift 4 support<br>Dropped iOS8 compatibility
 * **0.5.4** Fixed bug where blur view would reveal hidden layer<br>Improved view controller lifecycle handling<br>Scroll views can now be used with gesture dismissal
 * **0.5.3** Fixed memory leak with custom view controllers<br>Added UI automation & snapshot tests
-* **0.5.2** Fixed image scaling for default view 
+* **0.5.2** Fixed image scaling for default view
 * **0.5.1** Introduced custom button height parameter<br>Reintroduced iOS8 compatibility
 * **0.5.0** Swift 3 compatibility / removed iOS8
 * **0.4.0** iOS 8 compatibility